refactor: minimise public API surface in coker/__init__.py
Remove internal implementation details from the public API:
- Tape, Tracer, scalar_types, OP, Noop, Element, normalise

Move get_projection to coker.algebra (imported as coker.algebra.get_projection).

Fix all internal modules that were importing removed names via `from coker
import ...` to use their canonical submodule paths instead:
- coker.algebra.kernel: Tape, Tracer, Function, scalar_types, normalise, Noop
- coker.algebra.ops: OP, Noop
- coker.algebra.dimensions: FunctionSpace

Public API is now: function, Function, Scalar, VectorSpace, FunctionSpace,
Dimension, SymbolicVector, zeros, if_then_else.
