+++ src/coker/__init__.py
@@ -1,11 +1,6 @@
 from coker.algebra.kernel import (
-    Tape,
-    Tracer,
     Function,
     function,
-    scalar_types,
-    normalise,
-    get_projection,
     if_then_else,
 )
 from coker.algebra.dimensions import (
@@ -13,9 +8,6 @@ from coker.algebra.dimensions import (
     VectorSpace,
     Scalar,
     FunctionSpace,
-    Element,
 )
-from coker.algebra.ops import OP, Noop
 from coker.algebra.tensor import SymbolicVector
-
 from coker.algebra.factories import zeros
