+++ src/coker/backends/coker/layers.py
@@ -1,4 +1,6 @@
-from coker import Dimension, OP, scalar_types
+from coker.algebra.dimensions import Dimension
+from coker.algebra.ops import OP
+from coker.algebra.kernel import scalar_types
 from coker.backends import get_backend_by_name
 from coker.backends.coker.sparse_tensor import dok_ndarray
 from coker.backends.coker.memory import MemorySpec
