+++ src/coker/backends/coker/op_impl.py
@@ -1,5 +1,5 @@
 import numpy as np
-from coker import OP
+from coker.algebra.ops import OP
 from coker.backends.coker.sparse_tensor import (
     dok_ndarray,
     tensor_vector_product,
